+++ CHANGELOG.markdown
@@ -12,6 +12,7 @@
 - Redesigned window API
 - Redesigned progress bar API
 - Redesigned color editor/picker API
+- Redesigned child window API (previously known as child frame)
 - Updated layout API
 - Renderer errors implement std::error::Error
 - Glium renderer re-exports imgui and glium